+++ source/common.js
@@ -1,5 +1,9 @@
 import path from 'path'
 
+import serialize from './tools/serialize-javascript'
+
+import { exists } from './helpers'
+
 // returns a stub for webpack-assets.json if it doesn't exist yet
 // (because node.js and webpack are being run in parallel in development mode)
 export function default_webpack_assets()
@@ -52,4 +56,30 @@ export function normalize_options(options)
 			throw new Error(`You must specify file extensions for assets of type "${asset_type}"`)
 		}
 	}
+}
+
+// returns a CommonJS modules source.
+export function to_javascript_module_source(source)
+{
+	// if the asset source wasn't found - return an empty CommonJS module
+	if (!exists(source))
+	{
+		return 'module.exports = undefined'
+	}
+
+	// if it's already a common js module source
+	if (typeof source === 'string' && is_a_module_declaration(source))
+	{
+		return source
+	}
+
+	// generate javascript module source code based on the `source` variable
+	return 'module.exports = ' + serialize(source)
+}
+
+// detect if it is a CommonJS module declaration
+function is_a_module_declaration(source)
+{
+	return source.indexOf('module.exports = ') === 0 ||
+		/\s+module\.exports = .+/.test(source)
 }